requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
six==1.16.0
sniffio==1.3.1
terminal==0.4.0
